Admin can remove route or guide

+++ MyAdventure/Controllers/RoutesController.cs
@@ -211,6 +211,11 @@
         //[HttpPost]
         public IActionResult Delete(int id)
         {
+            var routeExists = this.routeService.CheckIfRouteExist(id);
+            if (!routeExists)
+            {
+                return BadRequest();
+            }
             this.routeService.DeleteRoute(id);
 
             return this.RedirectToAction(nameof(All));
